+++ crates/crypto/src/header/file.rs
@@ -32,7 +32,7 @@
 use std::io::{Cursor, Read, Seek, SeekFrom, Write};
 
 use crate::{
-	crypto::stream::{Algorithm, StreamDecryption, StreamEncryption},
+	crypto::stream::{Algorithm},
 	error::Error,
 	primitives::{generate_nonce, MASTER_KEY_LEN},
 	Protected,
@@ -40,8 +40,8 @@ use crate::{
 
 use super::{
 	keyslot::Keyslot,
-	metadata::{Metadata, MetadataVersion},
-	preview_media::{PreviewMedia, PreviewMediaVersion},
+	metadata::{Metadata},
+	preview_media::{PreviewMedia},
 };
 
 /// These are used to quickly and easily identify Spacedrive-encrypted files
@@ -98,192 +98,6 @@ impl FileHeader {
 			keyslots,
 			metadata: None,
 			preview_media: None,
-		}
-	}
-
-	/// This should be used for creating a header metadata item.
-	///
-	/// It handles encrypting the master key and metadata.
-	///
-	/// You will need to provide the user's password, and a semi-universal salt for hashing the user's password. This allows for extremely fast decryption.
-	///
-	/// Metadata needs to be accessed switfly, so a key management system should handle the salt generation.
-	pub fn add_metadata<T>(
-		&mut self,
-		version: MetadataVersion,
-		algorithm: Algorithm,
-		master_key: &Protected<[u8; MASTER_KEY_LEN]>,
-		metadata: &T,
-	) -> Result<(), Error>
-	where
-		T: ?Sized + serde::Serialize,
-	{
-		let metadata_nonce = generate_nonce(algorithm);
-
-		let encrypted_metadata = StreamEncryption::encrypt_bytes(
-			master_key.clone(),
-			&metadata_nonce,
-			algorithm,
-			&serde_json::to_vec(metadata).map_err(|_| Error::MetadataDeSerialization)?,
-			&[],
-		)?;
-
-		let metadata = Metadata {
-			version,
-			algorithm,
-			metadata_nonce,
-			metadata: encrypted_metadata,
-		};
-
-		self.metadata = Some(metadata);
-
-		Ok(())
-	}
-
-	/// This should be used for creating a header preview media item.
-	///
-	/// This handles encrypting the master key and preview media.
-	///
-	/// You will need to provide the user's password, and a semi-universal salt for hashing the user's password. This allows for extremely fast decryption.
-	///
-	/// Preview media needs to be accessed switfly, so a key management system should handle the salt generation.
-	pub fn add_preview_media(
-		&mut self,
-		version: PreviewMediaVersion,
-		algorithm: Algorithm,
-		master_key: &Protected<[u8; MASTER_KEY_LEN]>,
-		media: &[u8],
-	) -> Result<(), Error> {
-		let media_nonce = generate_nonce(algorithm);
-
-		let encrypted_media = StreamEncryption::encrypt_bytes(
-			master_key.clone(),
-			&media_nonce,
-			algorithm,
-			media,
-			&[],
-		)?;
-
-		let pvm = PreviewMedia {
-			version,
-			algorithm,
-			media_nonce,
-			media: encrypted_media,
-		};
-
-		self.preview_media = Some(pvm);
-
-		Ok(())
-	}
-
-	/// This function should be used to retrieve the metadata for a file
-	///
-	/// All it requires is pre-hashed keys returned from the KeyManager
-	///
-	/// A deserialized data type will be returned from this function
-	pub fn decrypt_metadata_from_prehashed<T>(
-		&self,
-		hashed_keys: Vec<Protected<[u8; 32]>>,
-	) -> Result<T, Error>
-	where
-		T: serde::de::DeserializeOwned,
-	{
-		let master_key = self.decrypt_master_key_from_prehashed(hashed_keys)?;
-
-		// could be an expensive clone (a few MiB at most)
-		if let Some(metadata) = self.metadata.clone() {
-			let metadata = StreamDecryption::decrypt_bytes(
-				master_key,
-				&metadata.metadata_nonce,
-				metadata.algorithm,
-				&metadata.metadata,
-				&[],
-			)?;
-
-			serde_json::from_slice::<T>(&metadata).map_err(|_| Error::MetadataDeSerialization)
-		} else {
-			Err(Error::NoMetadata)
-		}
-	}
-
-	/// This function should be used to retrieve the metadata for a file
-	///
-	/// All it requires is a password. Hashing is handled for you.
-	///
-	/// A deserialized data type will be returned from this function
-	pub fn decrypt_metadata<T>(&self, password: Protected<Vec<u8>>) -> Result<T, Error>
-	where
-		T: serde::de::DeserializeOwned,
-	{
-		let master_key = self.decrypt_master_key(password)?;
-
-		// could be an expensive clone (a few MiB at most)
-		if let Some(metadata) = self.metadata.clone() {
-			let metadata = StreamDecryption::decrypt_bytes(
-				master_key,
-				&metadata.metadata_nonce,
-				metadata.algorithm,
-				&metadata.metadata,
-				&[],
-			)?;
-
-			serde_json::from_slice::<T>(&metadata).map_err(|_| Error::MetadataDeSerialization)
-		} else {
-			Err(Error::NoMetadata)
-		}
-	}
-
-	/// This function is what you'll want to use to get the preview media for a file
-	///
-	/// All it requires is pre-hashed keys returned from the key manager
-	///
-	/// Once provided, a `Vec<u8>` is returned that contains the preview media
-	pub fn decrypt_preview_media_from_prehashed(
-		&self,
-		hashed_keys: Vec<Protected<[u8; 32]>>,
-	) -> Result<Protected<Vec<u8>>, Error> {
-		let master_key = self.decrypt_master_key_from_prehashed(hashed_keys)?;
-
-		// could be an expensive clone (a few MiB at most)
-		if let Some(pvm) = self.preview_media.clone() {
-			let media = StreamDecryption::decrypt_bytes(
-				master_key,
-				&pvm.media_nonce,
-				pvm.algorithm,
-				&pvm.media,
-				&[],
-			)?;
-
-			Ok(media)
-		} else {
-			Err(Error::NoPreviewMedia)
-		}
-	}
-
-	/// This function is what you'll want to use to get the preview media for a file
-	///
-	/// All it requires is the user's password. Hashing is handled for you.
-	///
-	/// Once provided, a `Vec<u8>` is returned that contains the preview media
-	pub fn decrypt_preview_media(
-		&self,
-		password: Protected<Vec<u8>>,
-	) -> Result<Protected<Vec<u8>>, Error> {
-		let master_key = self.decrypt_master_key(password)?;
-
-		// could be an expensive clone (a few MiB at most)
-		if let Some(pvm) = self.preview_media.clone() {
-			let media = StreamDecryption::decrypt_bytes(
-				master_key,
-				&pvm.media_nonce,
-				pvm.algorithm,
-				&pvm.media,
-				&[],
-			)?;
-
-			Ok(media)
-		} else {
-			Err(Error::NoPreviewMedia)
 		}
 	}
